agent-utils/pyproject.toml
[project]
name = "agent-utils"
version = "0.0.1"
description = "agent-utils"
authors = [{name = "Mateusz Krainski", email = "mt.krainski@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2,<2.0.0",
    "typer>=0.15.0,<1.0.0",
]

[project.scripts]
agent-utils = "agent_utils.cli:app"

[dependency-groups]
dev = [
    "coverage>=7.13.4,<8.0.0",
    "pytest>=9.0.2,<10.0.0",
    "ruff>=0.15.1,<0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
source = ["agent_utils"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "D",    # pydocstyle (flake8-docstrings)
    "ERA",  # eradicate
    "S",    # flake8-bandit
]
ignore = ["D104", "D100", "D101", "D106"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S603", "S607", "D103"]
"agent_utils/git_commit.py" = ["S603", "S607"]
"agent_utils/git_push.py" = ["S603", "S607"]
"agent_utils/gh_pr_create.py" = ["S603", "S607"]
"agent_utils/gh_pr_fetch.py" = ["S603", "S607"]
"agent_utils/gh_pr_reply.py" = ["S603", "S607"]
"agent_utils/gh_pr_checks.py" = ["S603", "S607"]
"agent_utils/gh_pr_list.py" = ["S603", "S607"]
"agent_utils/gh_pr_view.py" = ["S603", "S607"]
"agent_utils/gh_pr_close.py" = ["S603", "S607"]

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.uv]
add-bounds = "major"
preview = true

[tool.poe]
envfile = ".env"

[tool.poe.executor]
type = "uv"

[tool.poe.tasks]
configure = "./scripts/configure.sh"
update = "./scripts/update.sh"
hooks-run = "pre-commit run --all-files"
git-commit = "python -m agent_utils.git_commit"
git-push = "python -m agent_utils.git_push"
gh-pr-create = "python -m agent_utils.gh_pr_create"
gh-pr-fetch = "python -m agent_utils.gh_pr_fetch"
gh-pr-reply = "python -m agent_utils.gh_pr_reply"
gh-pr-checks = "python -m agent_utils.gh_pr_checks"
gh-pr-close = "python -m agent_utils.gh_pr_close"

[tool.poe.tasks.test]
sequence = [
    { cmd = "coverage run -m pytest --showlocals" },
    { cmd = "coverage report" },
]

[tool.poe.tasks.lint]
sequence = [
    { cmd = "ruff check ." },
    { cmd = "ruff format --check ." },
]

[tool.poe.tasks.test-full]
sequence = [
    { ref = "test" },
]
